Handle exceptions and fix issue with the script trying to select options on the dropdown before the options are filled
We are now waiting until the amount of options inside the dropdown
becomes greater than 1. Both dropdowns start with a single blank option,
when you the site finishes loading or after selecting a "unidade", the
site dynamically loads a list of options and adds to the dropdown, if we
try to fetch the elements on the dropdown before the site completes
fetching this list we end up with an empty dictionary because there will
be only the blank option inside the drropdown.